+++ Samples/Share/ShareBroadcastExtension/SampleHandler.swift
@@ -1,0 +1,13 @@
+//
+//  SampleHandler.swift
+//  ShareBroadcastExtension
+//
+//  Created by George Sealy on 23/06/21.
+//
+
+import ReplayKit
+import BJNiOSBroadcastExtension
+
+class SampleHandler: BJNSampleHandler {
+
+}